path_route added
So, what this function does is, it takes as arguments a begin coördinate
(usually the location of the cube) and an end-coördinate (as far from
the player as possible?) and it returns a path, manhattan style. So it
can't move diagonally, but will randomly move in the right direction
until target has been reached.