Importing and exporting of 'maps'
Maps can be imported and exported. Code is rather suboptimal, but seems
to work good enough.

0 means no agent can be created on the square.
1 means an agent can be created on the square.
X means an agent occupies the square.

This way, we can save and edit clusters to our liking.

Do note the personal_space variable in create_cluster.

Also, there is a probability that due to removal of possible locations
to place the agents, the amount of possible locations runs out before
all agents are placed. the program deals with this by catching the
<<ValueError>>, printing in the console that not all agents were made,
and continue like nothing happened because it did a great job and this
way we can fill up a complete cluster by just asking it to add as much
agents as it can up to an arbitrary large number.